Amend Uberon editor SOP about "abbreviation" synonyms.
The Uberon editor SOP is still recommending that synonyms that are
abbreviations should be marked with the synonym type
http://purl.obolibrary.org/obo/uberon/core#ABBREVIATION.

We have phased out the use of this particular synonym type several
months (maybe years?) ago, in favour of the OMO-defined synonym type
OMO:0003000. So here we simply amend the SOP to recommend using the
OMO-defined type.