Remove xrefs to HsapDv/MmusDv/OlatDv.
The SSLSO project is now providing mappings to Uberon in the form of a
SSSOM mapping set. The xrefs maintained in Uberon are redundant with
those mappings, and they seem to be less maintained (several Uberon
xrefs currently point to obsoleted HsapDv or MmusDv terms).

We remove those cross-references entirely, so that SSLSO becomes the
sole source of truth for the mappings between Uberon and
HsapDv/MmusDv/OlatDv (as it already is for all the other *Dv
ontologies).