Import RO:0002012 (occurrent part of).
We need 'occurrent part of' as it a parent of FBdv's 'substage of'
(FBdv:00018001); without that relation, there is no link between
'substage of' and 'part of', which means that the property chain

  'part of' o 'in taxon' SubPropertyOf: 'in taxon'

cannot apply to 'substage of' relationships. This is turn means that
some relations between FBdv terms cannot be properly translated into
class expressions when generating Composite Metazoan.