Unlink 'late embryonic stage' from chordate-specific terms.
The pharyngula and neurula stages are both specific to chordates. We
explicitly mark them as such (updating the existing -- and very broad --
taxon restrictiom from 'Eumetazoa' to 'Chordata') and amend their
definition accordingly.

'late embryonic stage', on the other hand, should not be restricted to
chordates. Therefore, we change the existing relation

  'late embryonic stage' SubClassOf 'preceded by' some 'pharyngula stage'

to a GCI

  ('late embryonic stage' and ('occurs in' some Chordata))
    SubClassOf 'preceded by' some 'pharyngula stage'

closes #3478