Remove "uvea 'part of' some 'anterior segment of eyeball'" axiom.
The uvea is not entirely contained within the anterior segment of
eyeball. It extends dorsally relative to the lens. This is explicitly
acknowledged in Uberon, where one of the parts of the uvea, the 'optic
choroid', is said to 'overlap' the 'posterior segment of eyeball'.

Furthermore, we have another term 'anterior uvea', which reinforces the
idea that the entire uvea spans over more than just the anterior part of
the eye.

So here we remove the axiom stating that the uvea is part of the
anterior segment of eyeball.