Make 'future brain vesicle' a (material) anatomical entity.
'future brain vesicle' is currently classified as an (immaterial)
anatomical space.

This is not consisten with other parts of Uberon:

(1) 'future brain vesicle' is part of the 'ventricular system of central
    nervous system', which is explicitly defined as "a set of
    *structures* containing cerebrospinal fluid in the brain" (so this
    is about the structures that delineate the anatomical cavitites,
    not the cavities themselves).

(2) The brain ventricles (which are derived from the 'future brain
    vesicles') are themselves considered as material anatomical
    entities (again, referring to the structures that delineate the
    cavities rather than the actual cavities).

So here we re-classify 'future brain vesicle' as a 'multicellular
anatomical structure'.