Include seeAlso annotations in mapping-derived xrefs.
When generating the cross-references from the SSSOM-maintained mappings,
annotate them with any seeAlso annotations that may be associated with
the mappings.

The intention is that, if a mapping editor has a reference to back up a
mapping statement, that reference can be stored in the SSSOM seeAlso
field, and once there it should be preserved when the mapping is
transformed into a classic cross-reference.