Add explicit taxon constraint to 'insect leg'.
Add an explicit taxon constraint to 'insect leg', so that the term is
restricted to Insecta.

This is not strictly necessary as the term is, in fact, *already*
taxon-restricted to Insecta through the following axiom:

> 'develops from' some 'insect prothoracic leg disc'

and the following property chain:

> 'has developmental contribution from' o 'in taxon'
>    SubPropertyOf: 'in taxon'

But having an explicit taxon constraint on the term itself makes the
intention clearer (and is also something recommended in #3376).