Amend 'insect leg'.
Fix the existing definition of 'insect leg', which was a weird mix of
the FBbt definition and the GO definition.

Remove the relation "'develops from' some 'imaginal disc'", as it is not
true for all insects that the legs develop from imaginal discs -- in
some species, the larva may have its own legs, from which the adult legs
can develop (that is the case, for example, in Lepidoptera).

Also add a proper reference, and remove the editor note that said the
term was mostly intended for GO -- this may have been the reason the
term was introduced initially but I don't think this is true anymore.
The term is being used to annotate data, not only to serve in a GO
definition.