Amend 'insect prothoracic leg'.
Remove the relationship "'develops from' some 'insect prothoracic leg
disc'", as it it not true for all insects that legs develop from
imaginal discs.

Amend the logical definition to make it part of the 'insect thoracic
segment' rather than the 'insect adult thoracic segment', as legs may
not be adult-specific structures -- in some species the larvae may have
legs too.

Remove the SubClassOf axioms that are redundant with the
EquivalentClasses axiom.