Remove the dynamic XML catalog.
The "dynamic XML catalog" (catalog-dynamic.xml) is a remnant of some
old, Owwltools-dependent workflows that have been replaced when the
custom Makefile was overhauled in #3028.

This catalog is no longer needed by any existing workflow, so we can
remove it along with the (also unused) variables that referred to it,
the rule that triggered its generation, and the script that actually
generated it.

closes #3504