Fix duplicate synonym.
This commit merges two annotation assertion axioms that only differ by
their own annotations, into a single axiom.

This is semantically identical, but the problem with having two separate
axioms with two different annotations is that, in the OBO format, the
order in which the two axioms are written is not deterministic, so the
axioms will sometimes be written as

  synonym: "lower limb skin" EXACT [FMA:23102]
  synonym: "lower limb skin" EXACT [https://orcid.org/0000-0002-189-0473]

and sometimes as

  synonym: "lower limb skin" EXACT [https://orcid.org/0000-0002-189-0473]
  synonym: "lower limb skin" EXACT [FMA:23102]

As a result, and in the past year alone (May 2024-May 2025), those two
lines have swapped places at least 12 times.

This is because, when deciding the order in which to write the "synonym"
tags, the OBO writer does not take the qualifier values into account, so
both tags are considered "equal" and their sorting order is thus
indetermined.

Of course the proper way to fix that would be to update the OBO writer
in the OWL API to make sure all tags are always sorted
deterministically, but it the meantime, it is much simpler to just avoid
having two separate axioms for what is, in fact, the same information.