Use built-in ODK feature to release SSSOM set.
In ODK 1.6, the built-in feature to release a mapping set has been
fixed, so we can use it. This means, in particular, that we no longer
need a custom rule to deploy the release to GitHub, the standard
ODK-generated rule will work.

We also declare the uberon.sssom.tsv set to be a "merged" set, even
though we still override the rule that creates it, because the standard
rule does not take care of "orienting" the mappings (making sure that
all UBERON:* terms are on the subject side).