Use built-in ODK features to build components.
Components that are derived from templates or fetched from remote
locations can now be completely handled by the standard Makefile, we no
longer need custom rules for them.

The mapping-derived component still needs a custom rule however, because
we want a finer control of the axioms (and their annotations) that are
derived from the mappings.