Fix production of the life-stages-composite subset.
The new rule to produce the life-stages-composite subset was based on
the assumption that a "life_stage" subset exists in the source ontology.

In fact there is no such subset; in the original (OWLTools-based) rule,
the subset tags were forcefully inject by a SPARQL query which found all
subclasses of "life cycle stage" and tagged them with the "life_stage"
subset.

With the ODK plugin, there is no longer any need for such inefficient
workflows, we can make a subset based directly on a query.