Add more automatic reviewers.
Update the reviewers.yml workflow so that Chris Mungall is not the only
person asked to review changes affecting logical definitions.

This cannot be done simply by adding more reviewers to the
`requested_reviewers' API call though, because doing so would cause the
action to fail if the author of the PR is _any_ of the reviewers that
would be listed here.

Instead, we refactor the workflow so that the "Assign reviewer" step is
in a distinct job, in which we call the 'requested_reviewers' API
endpoint as many times as we have reviewers to assign, so that if one of
them happens to be the PR's author, only that specific call will fail.