Add infrastructure to allocate definitive IDs.
Add a helper rule in the custom Makefile to automatically replace
temporary IDs in the -edit file by definitive IDs.

Add a new GitHub Actions workflow to perform the ID replacement
operation (1) whenever something is pushed to the master branch and (2)
on demand on any other branch.

This supports the two methods envisioned in #3564 for dealing with
temporary IDs:

(1) keep using temporary IDs throughout the life of a PR, and replace
them on the master branch _after_ the PR has been merged; this happens
entirely automatically;

(2) use temporary IDs throughout most of the life of a PR, but replace
them just before merging the PR; this has to be manually triggered by
someone, ideally the same person who decides to merge the PR.