Use ODK 1.6 in diff.yml workflow.
The diff.yml workflow, which powers the “gogoeditdiff” feature, is
still using the ODK image version 1.5.4. This leads to a systematic
failure, because at least one of the rules invoked by this workflow
(`make uberon-base.owl`) is dependent upon the new ROBOT option
`--include-subproperties`, which is not available in the version of
ROBOT provided with ODK 1.5.4.

We update the workflow to ensure it uses version 1.6 of the ODK.

closes #3575