Improve the check for changes to logical definitions.
Replace the existing GA workflow that checks for changes to logical
definitions by an improved check that:

* detects changes to logical definitions independently of the format of
  the -edit file;
* can additionally assess the "scale" of the changes (how many logical
  definitions are altered, and what are the consequences in terms of
  number of inferred subclasses).

If changes to logical definitions are detected, the new action will
always post a report highlighting the consequences of the changes.

In addition, if the changes are "big" (more than 10 logical definitions
changed, or at least one change that causes more than 10 inferred
SubClassOf axioms to change), the PR will be blocked (by requesting
changes) and Chris Mungall and David Osumi-Sutherland will be assigned
as reviewers.

closes #3568