Fix the automatic ID allocation workflow.
The automatic ID allocation workflow currently cannot be used in fully
automated mode (after a PR has been merged), because it will always use
the default GITHUB_TOKEN, which is not allowed to push to the master
branch.

The action has a dedicated token (which should be allowed to push to the
master branch) specifically intended for the purpose of automatic ID
allocation, but for the push step to use it, the checkout step needs to
be configured _not_ to persist the checkout credentials, which is what
we do here.