Test: Move a tag out of place.
This commit is a test for the "allocate definitive IDs" GitHub Action.
It moves a relationship tag slightly out of a place within a Term frame,
which has no effect on the contents of said frame but should trigger the
action to reserialize the edit file to restore the canonical order of
tags.